requirements.txt
Django                        # Framework webowy    
djangorestframework            # Biblioteka do tworzenia REST API
Pillow                         # Obsługa i przetwarzanie obrazów
psycopg2-binary         # Sterownik PostgreSQL w wersji binarnej
gunicorn                       # Produkcyjny serwer WSGI
uvicorn                        # Lekki serwer ASGI wykorzystywany w środowisku dev
dj-database-url                # Pomaga w konfiguracji baz danych przez URL
djangorestframework-simplejwt  # Implementacja JWT dla DRF
django-cors-headers            # Obsługa nagłówków CORS w Django
drf-yasg                        # Generowanie dokumentacji Swagger/OpenAPI

drf-spectacular            # Alternatywa dla drf-yasg, generuje dokumentację OpenAPI

pytest                     # Narzędzie do testowania aplikacji

django-filter            # Filtry dla Django REST Framework
python-dateutil       # Obsługa dat i czasu

djangorestframework-gis  # Obsługa geograficznych danych w Django REST Framework

